Switch to single user table
