--- src/main/java/com/team11/cab/model/User.java
@@ -1,37 +1,0 @@
-package com.team11.cab.model;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "user_authentication")
-public class User {
-	
-	@Id
-	private String username;
-	private String password;
-	
-	@OneToOne
-	@JoinColumn(name = "username")
-	private Member member;
-
-	public String getUsername() {
-		return username;
-	}
-
-	public void setUsername(String username) {
-		this.username = username;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-}
